Re-added snapshot API docs.
Separated tag/tag_category and pool/pool_category models for consistency.
