+++ server/src/model/mod.rs
@@ -1,7 +1,9 @@
 pub mod comment;
 pub mod enums;
 pub mod pool;
+pub mod pool_category;
 pub mod post;
 pub mod snapshot;
 pub mod tag;
+pub mod tag_category;
 pub mod user;
